always check for errors in execution
Previously we would only check for errors when completed == true but failed executions have (somewhat confusingly) completed == false which led to a loop until timed out.
We now always check for errors regardless of status since if prompt is in history it is guaranteed to be completed in either successful or failed state.